# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/client-apps": {
      "get": {
        "operationId": "ClientAdminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "列出全部定制客户端与功能配置",
        "tags": [
          "v1 client administration"
        ]
      },
      "post": {
        "operationId": "ClientAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientAppDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "注册新的定制客户端，无需修改核心领域代码",
        "tags": [
          "v1 client administration"
        ]
      }
    },
    "/api/v1/admin/client-apps/legacy-route-usage": {
      "get": {
        "operationId": "ClientAdminController_legacyRouteUsage",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientApp",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询兼容路由真实调用量，作为下线门禁依据",
        "tags": [
          "v1 client administration"
        ]
      }
    },
    "/api/v1/admin/client-apps/{clientKey}/features/{featureKey}": {
      "post": {
        "operationId": "ClientAdminController_updateFeature",
        "parameters": [
          {
            "name": "clientKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featureKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientFeatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建或更新客户端功能开关与导航元数据",
        "tags": [
          "v1 client administration"
        ]
      }
    },
...
  },
  "info": {
    "title": "ZhiPath Platform API",
    "description": "智途 ZhiPath、CodeNova 与后续定制客户端共享平台 API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "client-app": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-App"
      }
    },
    "schemas": {
      "CreateClientAppDto": {
        "type": "object",
        "properties": {
          "clientKey": {
            "type": "string",
            "example": "campus-web"
          },
          "name": {
            "type": "string",
            "example": "校园定制版"
          },
          "allowedOrigins": {
            "example": [
              "https://campus.example.com"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "theme": {
            "type": "object"
          },
          "features": {
            "example": [
              "dashboard",
              "learning-paths"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "clientKey",
          "name",
          "allowedOrigins"
        ]
      },
      "UpdateClientFeatureDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "rolloutPercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "default": 100
          },
          "config": {
            "type": "object"
          }
        },
        "required": [
          "enabled"
        ]
      },
...
    }
  }
}